+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- use locale files from mex-model
+
 ### Security
 
 ## [1.0.1] - 2025-10-07